Changed data shell over to use prev sets for compare
The idea is that a real use case would not be checking the grid to see if it already pathed that tile. Instead it would be using a set to ask if it already pathed over the location. As in theory we would not be updating a massive data structure assuming it would be there shortly later.

 ICBM mod is a good example of this as it edits the world over a thread. So the data could change under it while threading. As well we can't store a large data set to know if it has already pathed. Instead we must use small data sets to know the answer.